v3/backend/requirements.txt
django 
djangorestframework
markdown
django-filter